Added ability for game to be won, deleted now redundant script
Added a variable that tracks the number of times a note has been deleted, as well as moved the failure counter from the fail script into the keypress script. This renders the fail scripts useless, so was deleted.

Each counter isn't tracking properly as the numbers are being reset when each note is deleted. A workaround for this is needed.