Notes now use arrays to appear at various times to match the beat
An array of arrays for each level's amount of seconds to wait between notes spawning has been created. In each level's array, different numbers are written to signify the number of seconds (/10) to wait between each note spawning, allowing guitar hero - esque button presses in time to the song, rather than just one note spawning every half second every time.

There is a glitch where the time between each note will be divided by ten as it should for half of the notes, but not the other half. This needs to be looked into.