Major update -see desc. Multiple features added
The minigame prefab now has a camera attatched which it switches to when activating, so the view is dead on

Moving -

The minigame has been moved far below the stage rather than in it so it can look better and won't be infront of the computer screen monitor within the game

Visual -

Added the hacking background as well as a black background behind this for a high - tech look. Also changed the letter colour to white for better visibility. Additionally added rings around the letters.

Timing -

Added Assets - added previously given visual assets for the minigame to unity and used the majority of them,

Added camera change for minigame -

Fixed the timing of letters for level 1, the game now almost perfectly syncs up with the audio.

To do: The game no longer destroys itself after losing, this must be fixed. The camera must also return to the player upon finishing the minigame.

The letter spawner and hitboxes need to be moved for a better user experience - letters should spawn offscreen and the successful hit and fail boxes should be on the far left of the screen.