Added scrolling background effect, altered object positioning
Added a paralax scrolling grid effect behind the notes to help the player keep track of the notes - currently in early stages. Notes at different heights appear to move faster or slower which should be changed.

Camera change now causes the game to appear over the main one which could be very useful. Will add a black background to hide this.

Changed the minigame's distance from the camera so notes now spawn offscreen and the minigame takes up the entire screen.

Slightly modified the timing of the first level to be correct. Will be able to create new levels soon