Major FMOD Update, bug fixes
The FMOD files have been changed to allow different sections of songs to be played from the start. This is done by changing the section value between 1 and 15. Going between 16 and 20 will play larger sections of the song made for longer levels.

A large bug has been fixed where the song would sometimes not play upon starting the minigame, as well as the background scroll being de-synced.