Update I created last week - large change in scripts and efficiency
The menu handler script is now based off of the spawner script, in order for the buttons in the menu to directly reference and communicate with the level value. A logic oversight is present where the level value constantly changes between its new value given by the buttons and the default value 0. If this can be fixed - menus will be working in full.