I FIXED IT!! Each button in menus now spawn different levels.
It took me days, but I finally fixed an error where the spawner script, which handles which levels are created, couldn't communicate with the script attatched to each button. The fix was having the button script send an int value to a new script attached to the actual canvas, which could be successfully read by the spawner script if given a tag.